*List of Thing for Boss*
1. Particles I will add using old existing ones if there are no new ones.
2. Sword hit: because of how the model parenting and each parts' transform is i cant add mesh colliders that moves with the animation closely,
so climbing to hit crystals is kind of impossible to set up, maybe use hookshot thing to get to crystal to hit it with sword? righ now
i just have it set up as bullet destroys exposed crystal.
3. i am finishing up phase 3 need to work on kamakazi.
4. animation problems: Phase1: couldnt figure out a way to import the leftleg damage, leftleg loop, animations w/o crashing unity, still need a rightleg damage animation.
  phase 2: Figured out why boss' right hand digs into the ground. think there is a minor difference between the idle to walk model that i was using and the
rest of the animations, so the hand was bugging out and digging into the ground. im switching to another model and that should fix it.
phase3: would be nice if we can have some sort of animation between 2 and 3. if time allows.

if the level is ready, we can test the boss fight tomorrow and make tweaks. shouldnt see any huge problems aside from the ones ive mentioned.